+++ src/pagegen/utility.py
@@ -77,10 +77,14 @@ def relative_path(path):
 
 	return path.replace(getcwd()+sep, '')
 
+
 def urlify(string):
 	''' Anything wich isn't alphanumeric, - or _ gets replaced with a - '''
-	url=string.lower()
-	return sub('[^/a-z0-9-_.]', '-', url)
+	url = string.lower()
+	url = sub('[^/a-z0-9-_.]', '-', url)
+	# Replace any double dashes
+	url = sub('--', '-', url)
+	return url
 
 
 def report_error(code, message):